Add translate() and rename icon file

+++ translations/update_ts.sh
@@ -1,4 +1,4 @@
-#! /bin/sh
+#!/bin/bash
 
 #***************************************************************************
 #*                                                                         *
@@ -24,31 +24,74 @@
 #*                                                                         *
 #***************************************************************************
 
+if [ -z  "$1" ]; then
+    echo "You need add command. Try use '-h' for help:"
+    echo "sh update_ts.sh -h"
+    exit 1
+fi
+
+if [ "$#" -gt "1" ]; then
+    echo "You need add one command only. Try use '-h' for help:"
+    echo "sh update_ts.sh -h"
+    exit 1
+fi
+
+if [ "$1"='-*' ]; then
+    sPar=`expr substr $1 2 1`
+fi
+
+if [ "$sPar" = "h" ]; then
+    echo "Script for preparing translations file for A2plus Workbench"
+    echo "Copyright (c) 2019 kbwbe"
+    echo ""
+    echo "Commands:"
+    echo "         -c    - create A2plus.ts file from *.ui file and from all *.py files"
+    echo "         -h    - this help"
+    echo "         -m    - merge all A2plus_*.ts files to A2plus_*.qm files"
+fi
+
+
 #==============================================================================
-# Script for preparing translations file of A2plus Workbench
+# Script for create translations file of A2plus Workbench
 #
 # The script has to be started within the A2plus/translations folder
 #==============================================================================
 
+if [ "$sPar" = "c" ]; then
+
 # 1) Scan UI files for strings
-echo "1. Scan UI files for strings"
+    echo "1. Scan UI files for strings"
     lupdate ../GuiA2p/Resources/ui/*.ui -ts uifiles.ts
 
 # 2) Scan .py files for strings
-echo "2. Scan .py files for strings"
+    echo "2. Scan .py files for strings"
     pylupdate5 ../*.py -ts pyfiles.ts -verbose
 
 # 3) Combine both scans above
-echo "3. Combine both scans above"
-    lconvert -i uifiles.ts pyfiles.ts -o A2plus.ts
+    echo "3. Combine both scans above"
+    lconvert -i uifiles.ts pyfiles.ts -o A2plus.ts -sort-contexts -no-obsolete -verbose
 
 # 4) Remove temporary files
-echo "4. Remove temporary files"
+    echo "4. Remove temporary files:"
     rm uifiles.ts
-echo "                          uifiles.ts"
+    echo "                          uifiles.ts"
     rm pyfiles.ts
-echo "                          pyfiles.ts"
+    echo "                          pyfiles.ts"
 
 # 5) Message for user
-echo "5. You have fresh A2plus.ts file now"
+    echo "5. You have fresh A2plus.ts file now"
+fi
+
+
+#==============================================================================
+# Script for merge all translations of A2plus Workbench
+#
+# The script has to be started within the A2plus/translations folder
+#==============================================================================
+
+if [ "$sPar" = "m" ]; then
+    lrelease *_*.ts
+    echo "You have fresh A2plus_*.qm files now"
+fi
+
